+++ h5rdmtoolbox/ld/user/attributes.py
@@ -10,6 +10,7 @@ from ontolutils.classes.thing import is_url
 from rdflib import RDF, DCTERMS, XSD, URIRef, Literal, Graph, BNode
 
 from h5rdmtoolbox.ld.utils import get_attr_dtype_as_xsd, get_obj_bnode
+from .utils import to_uriref
 from ..rdf import FileRDFManager, RDFManager
 
 # FALLBACK_PREDICATE = SCHEMA.additionalProperty
@@ -35,14 +36,14 @@ def _is_date_str(value: str):
         return None
 
 
-def process_file_attribute(parent_obj, name, data, graph, file_uri: Union[URIRef, BNode]):
+def process_file_attribute(parent_obj, name, data, graph, file_uri: Union[URIRef, BNode], blank_node_iri_base):
     rdf_user_type = FileRDFManager(parent_obj.attrs).type
     if rdf_user_type is not None:
         if isinstance(rdf_user_type, (list, np.ndarray)):
             for t in rdf_user_type:
-                graph.add((file_uri, RDF.type, URIRef(t)))
+                graph.add((file_uri, RDF.type, to_uriref(t, blank_node_iri_base)))
         else:
-            graph.add((file_uri, RDF.type, URIRef(rdf_user_type)))
+            graph.add((file_uri, RDF.type, to_uriref(rdf_user_type, blank_node_iri_base)))
     rdf_user_predicate = FileRDFManager(parent_obj.attrs).predicate[name]
 
     rdf_user_object = FileRDFManager(parent_obj.attrs).object[name]
@@ -57,16 +58,17 @@ def process_file_attribute(parent_obj, name, data, graph, file_uri: Union[URIRef
     if rdf_user_predicate:
         if rdf_user_object:
             if isinstance(rdf_user_object, Literal):
-                graph.add((file_uri, URIRef(rdf_user_predicate), rdf_user_object))
+                graph.add((file_uri, to_uriref(rdf_user_predicate, blank_node_iri_base), rdf_user_object))
             if isinstance(rdf_user_object, str) and rdf_user_object.startswith('http'):
-                graph.add((file_uri, URIRef(rdf_user_predicate), URIRef(rdf_user_object)))
+                graph.add((file_uri, to_uriref(rdf_user_predicate, blank_node_iri_base),
+                           to_uriref(rdf_user_object, blank_node_iri_base)))
             elif isinstance(rdf_user_object, list):
                 for ruo in rdf_user_object:
                     if isinstance(ruo, Literal):
-                        graph.add((file_uri, URIRef(rdf_user_predicate), ruo))
+                        graph.add((file_uri, to_uriref(rdf_user_predicate, blank_node_iri_base), ruo))
 
                     elif isinstance(ruo, str) and ruo.startswith('http'):
-                        graph.add((file_uri, URIRef(rdf_user_predicate), URIRef(ruo)))
+                        graph.add((file_uri, to_uriref(rdf_user_predicate, blank_node_iri_base), URIRef(ruo)))
 
                     elif isinstance(ruo, dict):
                         try:
@@ -76,7 +78,8 @@ def process_file_attribute(parent_obj, name, data, graph, file_uri: Union[URIRef
                             if len(_subjects) != 1:
                                 warnings.warn(f"Error parsing JSON-LD object for attribute. name={name}, data={data}. "
                                               f"Expected exactly one subject, found {len(_subjects)}")
-                            obj_graph.add((file_uri, URIRef(rdf_user_predicate), list(_subjects)[0]))
+                            obj_graph.add(
+                                (file_uri, to_uriref(rdf_user_predicate, blank_node_iri_base), to_uriref(list(_subjects)[0], blank_node_iri_base)))
                             graph += obj_graph
                         except Exception as e:
                             warnings.warn(
@@ -90,7 +93,8 @@ def process_file_attribute(parent_obj, name, data, graph, file_uri: Union[URIRef
                     if len(_subjects) != 1:
                         warnings.warn(f"Error parsing JSON-LD object for attribute. name={name}, data={data}. "
                                       f"Expected exactly one subject, found {len(_subjects)}")
-                    obj_graph.add((file_uri, URIRef(rdf_user_predicate), list(_subjects)[0]))
+                    obj_graph.add((file_uri, to_uriref(rdf_user_predicate, blank_node_iri_base),
+                                   to_uriref(list(_subjects)[0], blank_node_iri_base)))
                     graph += obj_graph
                 except Exception as e:
                     warnings.warn(
@@ -101,19 +105,19 @@ def process_file_attribute(parent_obj, name, data, graph, file_uri: Union[URIRef
                 date_value, xsd_type = date_info
                 graph.add(
                     (file_uri,
-                     URIRef(rdf_user_predicate),
+                     to_uriref(rdf_user_predicate, blank_node_iri_base),
                      Literal(date_value, datatype=xsd_type))
                 )
             elif is_url(data):
                 graph.add(
                     (file_uri,
-                     URIRef(rdf_user_predicate),
+                     to_uriref(rdf_user_predicate, blank_node_iri_base),
                      URIRef(data))
                 )
             else:
                 graph.add(
                     (file_uri,
-                     URIRef(rdf_user_predicate),
+                     to_uriref(rdf_user_predicate, blank_node_iri_base),
                      Literal(data, datatype=get_attr_dtype_as_xsd(data)))
                 )
     return graph
@@ -124,16 +128,16 @@ def process_attribute(parent_obj, name, data, graph, blank_node_iri_base):
 
     parent_uri = rdf_manager.subject
     if parent_uri:
-        parent_uri = URIRef(parent_uri)
+        parent_uri = to_uriref(parent_uri, blank_node_iri_base)
     else:
         parent_uri = get_obj_bnode(parent_obj, blank_node_iri_base)
     rdf_user_type = rdf_manager.type
     if rdf_user_type is not None:
         if isinstance(rdf_user_type, (list, np.ndarray)):
             for t in rdf_user_type:
-                graph.add((parent_uri, RDF.type, URIRef(t)))
+                graph.add((parent_uri, RDF.type, to_uriref(t, blank_node_iri_base)))
         else:
-            graph.add((parent_uri, RDF.type, URIRef(rdf_user_type)))
+            graph.add((parent_uri, RDF.type, to_uriref(rdf_user_type, blank_node_iri_base)))
     rdf_user_predicate = rdf_manager.predicate[name]
     rdf_user_object = rdf_manager.object[name]
     if rdf_user_object and not rdf_user_predicate:
@@ -146,23 +150,37 @@ def process_attribute(parent_obj, name, data, graph, blank_node_iri_base):
     def _add_to_graph(_rdf_user_object, _graph):
         if _rdf_user_object:
             if isinstance(_rdf_user_object, Literal):
-                _graph.add((parent_uri, URIRef(rdf_user_predicate), _rdf_user_object))
+                _graph.add((parent_uri, to_uriref(rdf_user_predicate, blank_node_iri_base), _rdf_user_object))
             elif isinstance(_rdf_user_object, str) and _rdf_user_object.startswith('http'):
-                _graph.add((parent_uri, URIRef(rdf_user_predicate), URIRef(_rdf_user_object)))
+                _graph.add((parent_uri, to_uriref(rdf_user_predicate, blank_node_iri_base),
+                            to_uriref(_rdf_user_object, blank_node_iri_base)))
             elif isinstance(_rdf_user_object, list):
                 for _itm in _rdf_user_object:
                     _graph = _add_to_graph(_itm, _graph)
             else:
                 if isinstance(_rdf_user_object, dict):
                     try:
-                        obj_graph = Graph().parse(data=json.loads(json.dumps(_rdf_user_object)),
+                        jsonld_dict = json.loads(json.dumps(_rdf_user_object))
+                        obj_graph = Graph().parse(data=jsonld_dict,
                                                   format="json-ld")
-                        # relate the obj_graph with the predicate:
-                        _subjects = set(obj_graph.subjects())
-                        if len(_subjects) != 1:
-                            warnings.warn(f"Error parsing JSON-LD object for attribute. name={name}, data={data}. "
-                                          f"Expected exactly one subject, found {len(_subjects)}")
-                        obj_graph.add((parent_uri, URIRef(rdf_user_predicate), list(_subjects)[0]))
+
+                        if "@graph" not in jsonld_dict:
+                            use_subject = jsonld_dict.get("@id")
+                            obj_graph.add((parent_uri, to_uriref(rdf_user_predicate, blank_node_iri_base),
+                                           to_uriref(use_subject, blank_node_iri_base)))
+                        else:
+                            _graph_entities = jsonld_dict["@graph"]
+                            if len(_graph_entities) == 1:
+                                use_subject = _graph_entities[0].get("@id")
+                                obj_graph.add((parent_uri, to_uriref(rdf_user_predicate, blank_node_iri_base),
+                                               to_uriref(use_subject, blank_node_iri_base)))
+                            else:
+                                for _entity in _graph_entities:
+                                    if RDF.type in _entity and _entity[RDF.type] == str(SCHEMA.PropertyValue):
+                                        use_subject = _entity.get("@id")
+                                        obj_graph.add((parent_uri, to_uriref(rdf_user_predicate, blank_node_iri_base),
+                                                       to_uriref(use_subject, blank_node_iri_base)))
+
                         _graph += obj_graph
                     except Exception as e:
                         warnings.warn(
@@ -173,19 +191,19 @@ def process_attribute(parent_obj, name, data, graph, blank_node_iri_base):
                 date_value, xsd_type = date_info
                 graph.add(
                     (parent_uri,
-                     URIRef(rdf_user_predicate),
+                     to_uriref(rdf_user_predicate, blank_node_iri_base),
                      Literal(date_value, datatype=xsd_type))
                 )
             elif is_url(data):
                 graph.add(
                     (parent_uri,
-                     URIRef(rdf_user_predicate),
+                     to_uriref(rdf_user_predicate, blank_node_iri_base),
                      URIRef(data))
                 )
             else:
                 graph.add(
                     (parent_uri,
-                     URIRef(rdf_user_predicate),
+                     to_uriref(rdf_user_predicate, blank_node_iri_base),
                      Literal(data, datatype=get_attr_dtype_as_xsd(data)))
                 )
         return _graph