+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 Log of changes in the versions
 
+## v2.5.2
+
+- fix passing json-ld as RDF object
+- support for blank nodes in contextual semantic metadata
+
 ## v2.5.1
 
 - hotfix Organization class